requirements.txt
numpy
scipy
pandas
sentence-transformers
spacy
networkx
python-louvain
pypdf
scikit-learn
pyyaml
tqdm
ollama
